requirements.txt
playwright
openai
gitpython
pygithub
markdown
ipywidgets
IPython
pytest
psutil